+++ R/ggslopegraph.R
@@ -2,12 +2,12 @@
 #' @title Create Slopegraph from a data frame using ggplot2
 #' @description Convert an R data frame (containing a panel dataset, where rows are observations and columns are time periods) into an Edward Tufte-inspired Slopegraph using ggplot2
 #' @param data An observation-by-period data.frame, with at least two columns. Missing values are allowed.
-#' @param title A character string specifying a title. Passed to \code{\link[ggplot2]{ggtitle}}.
+#' @param main A character string specifying a title. Passed to \code{\link[ggplot2]{ggtitle}}.
 #' @param xlim A two-element numeric vector specifying the y-axis limits.
 #' @param ylim A two-element numeric vector specifying the y-axis limits.
 #' @param xlab A character string specifying an x-axis label. Passed to \code{\link[ggplot2]{scale_x_continuous}}.
 #' @param ylab A character string specifying an y-axis label. Passed to \code{\link[ggplot2]{scale_y_continuous}}, or \code{\link[ggplot2]{scale_y_reverse}} if \code{yrev = TRUE}.
-#' @param labels The labels to use for the slopegraph periods. Default is \code{names(df)}.#' @param xlab A character string specifying an x-axis label. Passed to \code{\link[ggplot2]{scale_x_continuous}}.
+#' @param xlabels The labels to use for the slopegraph periods. Default is \code{names(data)}.
 #' @param xbreaks Passed to \code{breaks} in \code{\link[ggplot2]{scale_x_continuous}}.
 #' @param ybreaks Passed to \code{breaks} in \code{\link[ggplot2]{scale_y_continuous}}.
 #' @param yrev A logical indicating whether to use \code{\link[ggplot2]{scale_y_reverse}} rather than the default \code{\link[ggplot2]{scale_y_continuous}}.
@@ -15,30 +15,48 @@
 #' @param col.lines A vector of colors for the slopegraph lines. Default is \code{par('fg')}.
 #' @param col.lab A vector of colors for the observation labels. Default is \code{par('fg')}.
 #' @param col.num A vector of colors for the number values. Default is \code{par('fg')}.
+#' @param offset.x A small offset for \code{segments}, to be used when positioning the numeric values. Default is \code{.1}.
 #' @param cex.lab A numeric value indicating the size of row labels. Default is \code{3}. See \code{\link[ggplot2]{geom_text}}.
 #' @param cex.num A numeric value indicating the size of numeric labels. Default is \code{3}. See \code{\link[ggplot2]{geom_text}}.
 #' @param lwd A vector of line width values for the slopegraph lines.
 #' @return A \code{\link[ggplot2]{ggplot}} object.
 #' @examples
+#' require("ggplot2")
+#' ## Tufte's Cancer Graph (to the correct scale)
+#' data(cancer)
+#' ggslopegraph(cancer, col.lines = 'gray', 
+#'   xlabels = c('5 Year','10 Year','15 Year','20 Year'))
+#' 
+#' ## Tufte's GDP Graph
+#' data(gdp)
+#' ggslopegraph(gdp, col.line='gray', xlabels = c('1970','1979'), 
+#'     main = 'Current Receipts of Goverment\nas a Percentage of Gross Domestic Product') + 
+#'   theme_bw()
+#' 
 #' ## Ranking of U.S. State populations
 #' data(states)
-#' ggslopegraph(states, title = 'Relative Rank of U.S. State Populations, 1790-1870', yrev = TRUE)
+#' ggslopegraph(states, 
+#'   main = 'Relative Rank of U.S. State Populations, 1790-1870', 
+#'   yrev = TRUE)
 #' 
 #' cls <- rep("black", nrow(states))
 #' cls[rownames(states) == "South Carolina"] <- "red"
 #' cls[rownames(states) == "Tennessee"] <- "blue"
-#' ggslopegraph(states, title = 'Relative Rank of U.S. State Populations, 1790-1870', yrev = TRUE, col.lines = cls, col.lab = cls)
+#' ggslopegraph(states, main = 'Relative Rank of U.S. State Populations, 1790-1870', 
+#'              yrev = TRUE, col.lines = cls, col.lab = cls)
+#'
 #' @seealso For a base graphics version, use \code{\link{slopegraph}}.
 #' @import ggplot2
 #' @export
 ggslopegraph <- 
 function(data, 
-         title = as.character(substitute(data)), 
-         xlab = "", 
-         ylab = "", 
+         main = as.character(substitute(data)), 
          xlim = c(-1L,ncol(data)+2L), 
          ylim = range(data, na.rm = TRUE), 
-         xbreaks = NULL,
+         xlab = "", 
+         ylab = "", 
+         xlabels = names(data),
+         xbreaks = seq_along(xlabels),
          ybreaks = NULL,
          yrev = ylim[1] > ylim[2], 
          decimals = NULL,
@@ -46,6 +64,7 @@ function(data,
          col.lab = "black",
          col.num = "black",
          lwd = 0.5,
+         offset.x = 0.1,
          cex.lab = 3L,
          cex.num = 3L)
 {
@@ -56,9 +75,6 @@ function(data,
     to_draw <- segmentize(as.matrix(data))
     colnames(to_draw) <- c("row", "x1", "x2", "y1", "y2")
     to_draw <- as.data.frame(to_draw)
-    
-    xoffset <- 0.1
-    yoffset <- 0
     
     # check decimal formatting
     fmt <- if (is.null(decimals)) "%0.0f" else paste0("%0.", decimals, "f")
@@ -82,7 +98,10 @@ function(data,
     # draw
     g <- ggplot() + 
         # segments
-        geom_segment(aes(x = x1 + xoffset, y = y1 + yoffset, xend = x2 - xoffset, yend = y2 - yoffset), 
+        geom_segment(aes(x = x1 + offset.x, 
+                         y = ifelse(y1 == y2, y1, (y1+((y2-y1)*offset.x))), 
+                         xend = x2 - offset.x, 
+                         yend = ifelse(y1 == y2, y2, (y2-((y2-y1)*offset.x)))), 
                      col = col.lines,
                      data = to_draw, inherit.aes = FALSE) + guides(fill = FALSE) + 
         # x1 labels 
@@ -95,9 +114,9 @@ function(data,
                   size = cex.num, hjust = 0.5) +
         # x-axis labels
         scale_x_continuous(name = xlab, breaks = xbreaks, 
-                           labels = names(data), limits = xlim) +
+                           labels = xlabels, limits = xlim) +
         # title
-        ggtitle(title) +
+        ggtitle(main) +
         if (isTRUE(yrev)) {
             scale_y_reverse(name = ylab, breaks = ybreaks, labels = NULL, limits = rev(ylim))
         } else {